Removed Git keep files
Many of these are unnecessary since the folders they were in have
content.

As part of this we removed some 'mailer' folders, which revealed that
config/environment.rb had a mailer configuration block that needed
those folders. We don't need a mailer for this project so this commit
also removed that block.